+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ### 1.1.0 (2013-xx-xx)
 
   * [Canvas] added 'getColorAt' method 
+  * [RGBColor][BC Breaks] 
+    - changed 'setRGBColor' to 'setFromRGBColor'
+    - added    setFromArray,setFromValue,setFromHex methdos